Bosses work now
This shit took WAAAAAAAAAAY too long. I don't know what happened, but the final boss wasn't detecting that he was on floor 5, so I had to patch it and manually set him to register that he's at floor 5 so that everything else could work. It's ugly as fuck, but it actually functions as intended now with no warnings popping up.